quick solution to p45
This one was interesting in that it was a direct extension of p44. Once
I saw that triangle numbers were a superset of hexagonal numbers, the
search space became a lot smaller. Of note, I implemented `isPentagonal`
from scratch in a very different way from P44 (in which I generated a
set and checked for membership); I did this because I had no idea what
the upper bound of my search space was in this problem.

Cool things learned: the existence of properFraction! I may have used
this before, but I had completely forgotten about it. Very useful little
function.

Things to change:

maybe implement a comprehension as opposed to a recursive search? I
would like to understand the memory trade offs of this better