Add solution to P47
This was was pretty neat in that there was almost no thought involved;
the answer just sort of fell out. While this is not generalize to accept
an arbitrary number of consecutive integers, its pretty tidy.

I tested my idea on both examples, and then ammended the search to look
for 4 consecutive numbers. The answer pops out in about .9 seconds on my
13in macbook air from 2013.